chore: Reorganize widget config colors
