+++ mobile/src/modules/widget/utils/customize.core.ts
@@ -1,0 +1,67 @@
+// Copyright (C) 2024 - present, MissingCore
+// SPDX-License-Identifier: AGPL-3.0-only
+
+import Storage from "expo-sqlite/kv-store";
+import { createStore } from "zustand/vanilla";
+
+import { isRecord } from "~/utils/validation";
+import { isWidgetConfigSupported } from "./customize";
+import { DEFAULT_WIDGET_CONFIG } from "../constants/Config";
+import type { WidgetConfig } from "../types";
+
+/** Local cache of widget configs so we don't overload calling the database. */
+export const widgetConfigCache = createStore<Record<string, WidgetConfig>>()(
+  () => ({}),
+);
+
+export async function getWidgetConfig(
+  widgetConfigKey: string,
+): Promise<WidgetConfig> {
+  const baseConfig = { ...DEFAULT_WIDGET_CONFIG };
+  if (!isWidgetConfigSupported(widgetConfigKey)) return baseConfig;
+
+  // First check to see if it's cached.
+  const cachedConfig = widgetConfigCache.getState()[widgetConfigKey];
+  if (cachedConfig) return cachedConfig;
+
+  const config = await Storage.getItemAsync(widgetConfigKey);
+  if (config) {
+    try {
+      const storedConfig = JSON.parse(config);
+      // Verify we're storing an object.
+      if (isRecord<WidgetConfig>(storedConfig)) {
+        const formattedConfig = { ...baseConfig, ...storedConfig };
+        widgetConfigCache.setState({ [widgetConfigKey]: formattedConfig });
+        return formattedConfig;
+      }
+    } catch {}
+  }
+
+  widgetConfigCache.setState({ [widgetConfigKey]: baseConfig });
+  return baseConfig;
+}
+
+export async function deleteWidgetConfig(widgetConfigKey: string) {
+  if (!isWidgetConfigSupported(widgetConfigKey)) return;
+
+  //? We get the following error if we use the sync method:
+  //?   - "Error: Call to function 'NativeDatabase.prepareAsync' has been rejected."
+  //?
+  //? This method might get called on a previously deleted widget due to
+  //? weird behavior that appeared with New Architecture implementation.
+  await Storage.removeItemAsync(widgetConfigKey);
+}
+
+export async function updateWidgetConfig(
+  widgetConfigKey: string,
+  config: WidgetConfig,
+) {
+  if (!isWidgetConfigSupported(widgetConfigKey)) return;
+
+  try {
+    widgetConfigCache.setState({ [widgetConfigKey]: config });
+
+    const stringifiedConfig = JSON.stringify(config);
+    await Storage.setItemAsync(widgetConfigKey, stringifiedConfig);
+  } catch {}
+}
