+++ mobile/src/modules/widget/utils/update.tsx
@@ -3,7 +3,7 @@
 
 import { requestWidgetUpdate } from "react-native-android-widget";
 
-import { getWidgetConfig, getWidgetConfigKey } from "./customize";
+import { getWidgetConfig, getWidgetConfigKey } from "./customize.core";
 import type { WidgetName } from "../impl";
 import { nameToWidget } from "../impl";
 import type { PlayerWidgetData } from "../types";
